Add cursor capture to Screeney
Former-commit-id: d1d2e7f50a521765ead030c65ec788eadc73c1da

+++ Clowd/Utilities/ScreenUtil.cs
@@ -47,7 +47,7 @@ namespace Clowd.Utilities
                     iconX = cursorInfo.ptScreenPos.x - ((int)iconInfo.xHotspot) - origin.X;
                     iconY = cursorInfo.ptScreenPos.y - ((int)iconInfo.yHotspot) - origin.Y;
 
-                    // Is this a monochrome cursor?
+                    // Is this a color cursor or a monochrome one?
                     if (iconInfo.hbmColor != IntPtr.Zero)
                     {
                         using (Icon curIcon = Icon.FromHandle(hicon))
@@ -71,8 +71,8 @@ namespace Clowd.Utilities
 
                             var resultHdc = g.GetHdc();
                             var size = maskBitmap.Width;
-                            GDI32.BitBlt(resultHdc, iconX, iconY, size, size, maskHdc, 0, 0, (int)TernaryRasterOperations.SRCAND);
-                            GDI32.BitBlt(resultHdc, iconX, iconY, size, size, maskHdc, 0, size, (int)TernaryRasterOperations.SRCINVERT);
+                            GDI32.BitBlt(resultHdc, iconX, iconY, size, size, maskHdc, 0, 0, TernaryRasterOperations.SRCAND);
+                            GDI32.BitBlt(resultHdc, iconX, iconY, size, size, maskHdc, 0, size, TernaryRasterOperations.SRCINVERT);
                             g.ReleaseHdc(resultHdc);
 
                             IntPtr newPtr = GDI32.SelectObject(maskHdc, oldPtr);
